feat: a download asks for its connections, up to 256, and the row remembers the count
Connections::auto is the engine's judgement and Connections::fixed exactly so many; the row
carries the count into the database, a version 2 column a version 1 file gains on open, so
a resume after a restart opens what was asked for. The speed limit and the default count
are preferences, and the limit reaches the engine at launch.

+++ src/engine/segments.rs
@@ -72,7 +72,7 @@ impl Plan {
 	/// The span cut into `parts` equal pieces, each at least `min_segment` long; fewer pieces
 	/// when the span is too short for that many. What a download starts with when it is allowed
 	/// several connections from the outset.
-	pub fn split(span: Span, parts: u8, min_segment: u64) -> Plan {
+	pub fn split(span: Span, parts: u16, min_segment: u64) -> Plan {
 		let min_segment = min_segment.max(1);
 		let parts = (u64::from(parts.max(1))).min(span.len() / min_segment).max(1);
 		let each = span.len() / parts;
